添加子弹击中水面的特效

+++ src/main/java/com/atsuishio/superbwarfare/client/particle/CustomCloudOption.java
@@ -20,7 +20,7 @@ public class CustomCloudOption implements ParticleOptions {
                     Codec.FLOAT.fieldOf("g").forGetter(option -> option.green),
                     Codec.FLOAT.fieldOf("b").forGetter(option -> option.blue),
                     Codec.INT.fieldOf("life").forGetter(option -> option.life),
-                    Codec.INT.fieldOf("size").forGetter(option -> option.size),
+                    Codec.FLOAT.fieldOf("size").forGetter(option -> option.size),
                     Codec.BOOL.fieldOf("cooldown").forGetter(option -> option.cooldown),
                     Codec.BOOL.fieldOf("light").forGetter(option -> option.light)
             ).apply(builder, CustomCloudOption::new));
@@ -28,7 +28,7 @@ public class CustomCloudOption implements ParticleOptions {
     public static final StreamCodec<ByteBuf, CustomCloudOption> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.VECTOR3F, CustomCloudOption::getColor,
             ByteBufCodecs.INT, CustomCloudOption::getLife,
-            ByteBufCodecs.INT, CustomCloudOption::getSize,
+            ByteBufCodecs.FLOAT, CustomCloudOption::getSize,
             ByteBufCodecs.BOOL, CustomCloudOption::getCooldown,
             ByteBufCodecs.BOOL, CustomCloudOption::getLight,
             CustomCloudOption::new
@@ -38,16 +38,16 @@ public class CustomCloudOption implements ParticleOptions {
     private final float green;
     private final float blue;
     private final int life;
-    private final int size;
+    private final float size;
     private final boolean cooldown;
     private final boolean light;
 
 
-    public CustomCloudOption(Vector3f color, int life, int size, boolean cooldown, boolean light) {
+    public CustomCloudOption(Vector3f color, int life, float size, boolean cooldown, boolean light) {
         this(color.x, color.y, color.z, life, size, cooldown, light);
     }
 
-    public CustomCloudOption(float r, float g, float b, int life, int size, boolean cooldown, boolean light) {
+    public CustomCloudOption(float r, float g, float b, int life, float size, boolean cooldown, boolean light) {
         this.red = r;
         this.green = g;
         this.blue = b;
@@ -77,7 +77,7 @@ public class CustomCloudOption implements ParticleOptions {
         return life;
     }
 
-    public int getSize() {
+    public float getSize() {
         return size;
     }
